+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - The `corpus_config_url` setting is replaced by `get_corpus_ids`, see [doc/frontend_devel.md](./doc/frontend_devel.md)
 - The `structService` service is replaced by non-AngularJS async functions in `@/backend/attr-values`
+- The `lexicons` service is replaced by non-AngularJS async functions in `@/backend/lexicons`
 - The `httpConfAddMethod*` util functions were refactored:
   - The `$.ajax` case of `httpConfAddMethod` was extracted into `ajaxConfAddMethod`
   - `httpConfAddMethodFetch` was renamed to `fetchConfAddMethod`